+++ nuntly-java-core/src/main/kotlin/com/nuntly/models/domains/DomainUpdateParams.kt
@@ -281,6 +281,7 @@ private constructor(
     override fun _queryParams(): QueryParams = additionalQueryParams
 
     class Body
+    @JsonCreator(mode = JsonCreator.Mode.DISABLED)
     private constructor(
         private val clickTracking: JsonField<Boolean>,
         private val openTracking: JsonField<Boolean>,