+++ nuntly-java-core/src/main/kotlin/com/nuntly/models/emails/EmailCancelResponse.kt
@@ -18,6 +18,7 @@ import java.util.Objects
 import kotlin.jvm.optionals.getOrNull
 
 class EmailCancelResponse
+@JsonCreator(mode = JsonCreator.Mode.DISABLED)
 private constructor(
     private val id: JsonField<String>,
     private val kind: JsonField<Kind>,