+++ nuntly-java-core/src/main/kotlin/com/nuntly/models/emails/EmailListPageResponse.kt
@@ -19,6 +19,7 @@ import java.util.Optional
 import kotlin.jvm.optionals.getOrNull
 
 class EmailListPageResponse
+@JsonCreator(mode = JsonCreator.Mode.DISABLED)
 private constructor(
     private val data: JsonField<List<EmailListResponse>>,
     private val nextCursor: JsonField<String>,