+++ nuntly-java-core/src/main/kotlin/com/nuntly/models/emails/events/EventListResponse.kt
@@ -20,6 +20,7 @@ import java.util.Objects
 import kotlin.jvm.optionals.getOrNull
 
 class EventListResponse
+@JsonCreator(mode = JsonCreator.Mode.DISABLED)
 private constructor(
     private val id: JsonField<String>,
     private val details: JsonField<Details>,