+++ nuntly-java-core/src/main/kotlin/com/nuntly/models/webhooks/events/EventListResponse.kt
@@ -30,6 +30,7 @@ import java.util.Objects
 import kotlin.jvm.optionals.getOrNull
 
 class EventListResponse
+@JsonCreator(mode = JsonCreator.Mode.DISABLED)
 private constructor(
     private val id: JsonField<String>,
     private val data: JsonField<Event>,