+++ nuntly-java-core/src/test/kotlin/com/nuntly/models/webhooks/events/EventReplayParamsTest.kt
@@ -5,17 +5,20 @@ package com.nuntly.models.webhooks.events
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-internal class EventRetryParamsTest {
+internal class EventReplayParamsTest {
 
     @Test
     fun create() {
-        EventRetryParams.builder().id("wh_YNtYn86oYZmP1ZHbnUBvXXFt").eventId("event_id").build()
+        EventReplayParams.builder().id("wh_YNtYn86oYZmP1ZHbnUBvXXFt").eventId("event_id").build()
     }
 
     @Test
     fun pathParams() {
         val params =
-            EventRetryParams.builder().id("wh_YNtYn86oYZmP1ZHbnUBvXXFt").eventId("event_id").build()
+            EventReplayParams.builder()
+                .id("wh_YNtYn86oYZmP1ZHbnUBvXXFt")
+                .eventId("event_id")
+                .build()
 
         assertThat(params._pathParam(0)).isEqualTo("wh_YNtYn86oYZmP1ZHbnUBvXXFt")
         assertThat(params._pathParam(1)).isEqualTo("event_id")